requirements.txt
numpy==2.2.6
pandas==2.3.3
seaborn==0.13.2
matplotlib==3.10.8
torch==2.10.0
scikit-learn==1.7.2
spacy==3.8.11
shap==0.49.1
plotly==6.6.0
umap-learn==0.5.11
tqdm==4.67.3
streamlit==1.55.0
python-dotenv==1.2.2
pydantic==2.12.5
ipykernel==7.2.0
jupyter==1.1.1
scipy==1.15.3